feat(graphics): add new collage class with functionality
This commit includes the implementation of a new Collage class within the graphics scope. This class features necessary attributes and methods for creating, labeling, and managing collages of images.

refactor(graphics): replace _collage.py with collage.py

The old _collage.py file has been deleted and replaced with the new collage.py file, which provides more comprehensive and efficient functionality for handling image collages.

Note: The commit message could vary based on specific project's commit conventions and the level of detail desired. The provided message is a general overview assuming conventional commit message style and changes made in the diff.
